requirements.txt
flask==3.1.2
beautifulsoup4==4.14.2
flask-sqlalchemy==3.1.1
flask-login==0.6.3
python-dotenv==1.2.1
pymysql==1.1.2
cryptography==46.0.3
mysql-connector-python==9.5.0